requirements.txt
flask>=3.0.0
browser-use==0.12.6
langchain-anthropic>=0.3.0
langchain>=0.3.0
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
requests>=2.32.0
slack-bolt>=1.21.0
pydantic>=2.10.0
playwright>=1.40.0
